Decompose mesh-llm-system modules
