+++ crates/mesh-llm-system/src/autoupdate/bundle_install.rs
@@ -1,0 +1,531 @@
+async fn install_latest_bundle(
+    exe: &Path,
+    install_dir: &Path,
+    release: &ReleaseInfo,
+    asset_name: &str,
+    expected_flavor: backend::BinaryFlavor,
+    action: PostInstallAction,
+) -> Result<InstallOutcome> {
+    let unique = format!(
+        "{}-{}",
+        std::process::id(),
+        std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .unwrap_or_default()
+            .as_secs()
+    );
+    let workspace = install_dir.join(format!(".mesh-llm-update-{unique}"));
+    let extracted = workspace.join("bundle");
+    let archive = workspace.join(asset_name);
+    let backup = workspace.join("backup");
+
+    std::fs::create_dir_all(&extracted)
+        .with_context(|| format!("Failed to create update workspace {}", workspace.display()))?;
+
+    let result = async {
+        download_url(&release_asset_url(&release.tag, asset_name), &archive).await?;
+        extract_bundle_archive(&archive, &extracted)?;
+        let staged_files = collect_bundle_files(&extracted, expected_flavor)?;
+        finish_bundle_install(
+            exe,
+            install_dir,
+            &workspace,
+            &extracted,
+            &backup,
+            &staged_files,
+            action,
+        )?;
+        Ok::<InstallOutcome, anyhow::Error>(install_outcome(action))
+    }
+    .await;
+
+    if !matches!(result, Ok(InstallOutcome::HandoffAndExit)) {
+        let _ = std::fs::remove_dir_all(&workspace);
+    }
+    result
+}
+
+async fn download_url(url: &str, path: &Path) -> Result<()> {
+    let response = reqwest::Client::builder()
+        .timeout(std::time::Duration::from_secs(300))
+        .build()
+        .context("Build release download HTTP client")?
+        .get(url)
+        .header("User-Agent", "mesh-llm")
+        .send()
+        .await
+        .with_context(|| format!("Download release asset {url}"))?
+        .error_for_status()
+        .with_context(|| format!("Release asset request failed for {url}"))?;
+    let bytes = response
+        .bytes()
+        .await
+        .with_context(|| format!("Read release asset body from {url}"))?;
+    std::fs::write(path, &bytes)
+        .with_context(|| format!("Write release asset {}", path.display()))?;
+    Ok(())
+}
+
+fn extract_bundle_archive(archive: &Path, extracted: &Path) -> Result<()> {
+    match archive
+        .extension()
+        .and_then(|ext| ext.to_str())
+        .map(|ext| ext.to_ascii_lowercase())
+        .as_deref()
+    {
+        Some("zip") => extract_zip_archive(archive, extracted),
+        _ => extract_tar_archive(archive, extracted),
+    }
+}
+
+fn extract_tar_archive(archive: &Path, extracted: &Path) -> Result<()> {
+    let status = std::process::Command::new("tar")
+        .arg("-xzf")
+        .arg(archive)
+        .arg("-C")
+        .arg(extracted)
+        .arg("--strip-components=1")
+        .status()
+        .with_context(|| format!("Failed to extract {}", archive.display()))?;
+    anyhow::ensure!(status.success(), "tar extraction failed");
+    Ok(())
+}
+
+fn extract_zip_archive(archive: &Path, extracted: &Path) -> Result<()> {
+    let file = std::fs::File::open(archive)
+        .with_context(|| format!("Failed to open {}", archive.display()))?;
+    let mut zip = zip::ZipArchive::new(file)
+        .with_context(|| format!("Failed to read ZIP archive {}", archive.display()))?;
+
+    for index in 0..zip.len() {
+        let mut entry = zip.by_index(index)?;
+        let enclosed = entry
+            .enclosed_name()
+            .context("ZIP archive contained an invalid path")?;
+        let mut components = enclosed.components();
+        let _top_level = components.next();
+        let relative: PathBuf = components.collect();
+        if relative.as_os_str().is_empty() {
+            continue;
+        }
+
+        let output = extracted.join(&relative);
+        if entry.is_dir() {
+            std::fs::create_dir_all(&output)
+                .with_context(|| format!("Failed to create {}", output.display()))?;
+            continue;
+        }
+
+        if let Some(parent) = output.parent() {
+            std::fs::create_dir_all(parent)
+                .with_context(|| format!("Failed to create {}", parent.display()))?;
+        }
+        let mut out = std::fs::File::create(&output)
+            .with_context(|| format!("Failed to create {}", output.display()))?;
+        io::copy(&mut entry, &mut out)
+            .with_context(|| format!("Failed to extract {}", output.display()))?;
+    }
+
+    Ok(())
+}
+
+#[cfg(test)]
+mod zip_tests {
+    use super::*;
+    use std::fs;
+    use std::io::Write;
+    use std::time::{SystemTime, UNIX_EPOCH};
+
+    use zip::write::SimpleFileOptions;
+    use zip::CompressionMethod;
+
+    fn unique_temp_dir(prefix: &str) -> PathBuf {
+        let nanos = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .as_nanos();
+        let dir = std::env::temp_dir().join(format!("{}_{}", prefix, nanos));
+        // Best-effort cleanup in case something is left behind from a previous run.
+        let _ = fs::remove_dir_all(&dir);
+        fs::create_dir_all(&dir).expect("failed to create temporary directory");
+        dir
+    }
+
+    #[test]
+    fn extract_zip_archive_strips_top_level_directory() -> Result<()> {
+        let base_dir = unique_temp_dir("mesh_llm_extract_zip_test");
+        let archive_path = base_dir.join("bundle.zip");
+        let extracted_dir = base_dir.join("extracted");
+        fs::create_dir_all(&extracted_dir)?;
+
+        // Create a ZIP with a single top-level directory, similar to the release packager.
+        let file = std::fs::File::create(&archive_path)
+            .with_context(|| format!("Failed to create test archive {}", archive_path.display()))?;
+        let mut writer = zip::ZipWriter::new(file);
+        let options = SimpleFileOptions::default().compression_method(CompressionMethod::Stored);
+
+        // Top-level directory.
+        writer.add_directory("bundle-1.0.0/", options)?;
+        // Nested directory and file under the top-level directory.
+        writer.add_directory("bundle-1.0.0/bin/", options)?;
+        writer.start_file("bundle-1.0.0/bin/server", options)?;
+        writer.write_all(b"dummy-server")?;
+
+        writer
+            .finish()
+            .with_context(|| "Failed to finalize test ZIP archive")?;
+
+        // Now extract and verify that the top-level directory is stripped.
+        extract_zip_archive(&archive_path, &extracted_dir)?;
+
+        let server_path = extracted_dir.join("bin").join("server");
+        anyhow::ensure!(
+            server_path.is_file(),
+            "Expected extracted server file at {}",
+            server_path.display()
+        );
+
+        let top_level = extracted_dir.join("bundle-1.0.0");
+        anyhow::ensure!(
+            !top_level.exists(),
+            "Top-level directory should have been stripped, but {} exists",
+            top_level.display()
+        );
+
+        Ok(())
+    }
+}
+fn collect_bundle_files(
+    extracted: &Path,
+    expected_flavor: backend::BinaryFlavor,
+) -> Result<Vec<String>> {
+    let _ = expected_flavor;
+
+    let mut files = Vec::new();
+    for entry in std::fs::read_dir(extracted)
+        .with_context(|| format!("Failed to read {}", extracted.display()))?
+    {
+        let entry = entry?;
+        let file_type = entry.file_type()?;
+        let name = entry.file_name();
+        let name = name.to_string_lossy().to_string();
+        if file_type.is_dir() {
+            anyhow::bail!("Unexpected directory in bundle: {name}");
+        }
+        if file_type.is_file() {
+            files.push(name);
+        }
+    }
+
+    anyhow::ensure!(!files.is_empty(), "Downloaded bundle was empty");
+    anyhow::ensure!(
+        files.iter().any(|name| name == &mesh_binary_name()),
+        "Downloaded bundle missing {}",
+        mesh_binary_name()
+    );
+    files.sort_by_key(|name| (name == &mesh_binary_name(), name.clone()));
+    Ok(files)
+}
+
+#[cfg(not(windows))]
+fn backup_existing_file(
+    install_dir: &Path,
+    backup: &Path,
+    name: &str,
+    backed_up: &mut Vec<String>,
+) -> Result<()> {
+    if backed_up.iter().any(|existing| existing == name) {
+        return Ok(());
+    }
+
+    let dest = install_dir.join(name);
+    if !dest.exists() {
+        return Ok(());
+    }
+
+    let backup_path = backup.join(name);
+    std::fs::rename(&dest, &backup_path).with_context(|| {
+        format!(
+            "Failed to move {} to {}",
+            dest.display(),
+            backup_path.display()
+        )
+    })?;
+    backed_up.push(name.to_string());
+    Ok(())
+}
+
+#[cfg(not(windows))]
+fn replace_bundle_files(
+    install_dir: &Path,
+    extracted: &Path,
+    backup: &Path,
+    staged_files: &[String],
+) -> Result<()> {
+    use std::collections::BTreeSet;
+
+    std::fs::create_dir_all(backup)
+        .with_context(|| format!("Failed to create backup dir {}", backup.display()))?;
+
+    let managed_names: BTreeSet<String> = BTreeSet::from([mesh_binary_name()]);
+
+    let mut backed_up = Vec::new();
+    for name in managed_names {
+        backup_existing_file(install_dir, backup, &name, &mut backed_up)?;
+    }
+    for name in staged_files {
+        backup_existing_file(install_dir, backup, name, &mut backed_up)?;
+    }
+
+    let mut installed = Vec::new();
+    for name in staged_files {
+        let source = extracted.join(name);
+        let dest = install_dir.join(name);
+        if let Err(err) = std::fs::rename(&source, &dest) {
+            rollback_bundle_replace(install_dir, backup, &installed, &backed_up);
+            return Err(err).with_context(|| {
+                format!(
+                    "Failed to install {} into {}",
+                    source.display(),
+                    dest.display()
+                )
+            });
+        }
+        installed.push(name.clone());
+    }
+
+    Ok(())
+}
+
+#[cfg(not(windows))]
+fn install_outcome(action: PostInstallAction) -> InstallOutcome {
+    match action {
+        PostInstallAction::RestartCurrentProcess => InstallOutcome::RestartNow,
+        PostInstallAction::ExitAfterInstall => InstallOutcome::ExitNow,
+    }
+}
+
+#[cfg(windows)]
+fn install_outcome(_action: PostInstallAction) -> InstallOutcome {
+    InstallOutcome::HandoffAndExit
+}
+
+#[cfg(not(windows))]
+fn finish_bundle_install(
+    _exe: &Path,
+    install_dir: &Path,
+    _workspace: &Path,
+    extracted: &Path,
+    backup: &Path,
+    staged_files: &[String],
+    _action: PostInstallAction,
+) -> Result<()> {
+    replace_bundle_files(install_dir, extracted, backup, staged_files)
+}
+
+#[cfg(windows)]
+fn finish_bundle_install(
+    exe: &Path,
+    install_dir: &Path,
+    workspace: &Path,
+    extracted: &Path,
+    backup: &Path,
+    staged_files: &[String],
+    action: PostInstallAction,
+) -> Result<()> {
+    use std::process::Command;
+
+    let script = workspace.join("apply-update.ps1");
+    let script_body = windows_update_script(
+        exe,
+        install_dir,
+        workspace,
+        extracted,
+        backup,
+        staged_files,
+        action,
+    )?;
+    std::fs::write(&script, script_body)
+        .with_context(|| format!("Failed to write {}", script.display()))?;
+
+    Command::new("powershell")
+        .arg("-NoProfile")
+        .arg("-ExecutionPolicy")
+        .arg("Bypass")
+        .arg("-File")
+        .arg(&script)
+        .spawn()
+        .with_context(|| format!("Failed to launch Windows updater {}", script.display()))?;
+
+    Ok(())
+}
+
+#[cfg(windows)]
+fn windows_update_script(
+    exe: &Path,
+    install_dir: &Path,
+    workspace: &Path,
+    extracted: &Path,
+    backup: &Path,
+    staged_files: &[String],
+    action: PostInstallAction,
+) -> Result<String> {
+    use std::collections::BTreeSet;
+
+    let staged_json = serde_json::to_string(staged_files)?;
+    let restart_after_update = matches!(action, PostInstallAction::RestartCurrentProcess);
+    let args: Vec<String> = std::env::args_os()
+        .skip(1)
+        .map(|arg| arg.to_string_lossy().to_string())
+        .collect();
+    let args_json = serde_json::to_string(&args)?;
+
+    let managed_names: BTreeSet<String> = BTreeSet::from([mesh_binary_name()]);
+    let managed_json = serde_json::to_string(&managed_names.into_iter().collect::<Vec<_>>())?;
+
+    let quote = |path: &Path| path.to_string_lossy().replace('\'', "''");
+
+    let args_json_ps = args_json.replace('\'', "''");
+    let managed_json_ps = managed_json.replace('\'', "''");
+    let staged_json_ps = staged_json.replace('\'', "''");
+
+    Ok(format!(
+        r#"$ErrorActionPreference = 'Stop'
+$installDir = '{install_dir}'
+$workspace = '{workspace}'
+$stagingDir = '{staging_dir}'
+$backupDir = '{backup_dir}'
+$exePath = '{exe_path}'
+$waitPid = {wait_pid}
+$restartAfterUpdate = {restart_after_update}
+$managedNames = @((ConvertFrom-Json '{managed_json_ps}'))
+$stagedNames = @((ConvertFrom-Json '{staged_json_ps}'))
+$args = @((ConvertFrom-Json '{args_json_ps}'))
+
+function Restore-Backups([string[]]$BackedUpNames, [string[]]$InstalledNames) {{
+    foreach ($name in $InstalledNames) {{
+        $dest = Join-Path $installDir $name
+        Remove-Item $dest -Force -ErrorAction SilentlyContinue
+    }}
+    foreach ($name in $BackedUpNames) {{
+        $backupPath = Join-Path $backupDir $name
+        $dest = Join-Path $installDir $name
+        if (Test-Path $backupPath) {{
+            Move-Item -Force $backupPath $dest
+        }}
+    }}
+}}
+
+while (Get-Process -Id $waitPid -ErrorAction SilentlyContinue) {{
+    Start-Sleep -Milliseconds 200
+}}
+
+$backedUp = New-Object System.Collections.Generic.List[string]
+$installed = New-Object System.Collections.Generic.List[string]
+
+try {{
+    New-Item -ItemType Directory -Path $backupDir -Force | Out-Null
+
+    foreach ($name in $managedNames) {{
+        if ($stagedNames -contains $name) {{
+            continue
+        }}
+        $dest = Join-Path $installDir $name
+        if (-not (Test-Path $dest)) {{
+            continue
+        }}
+        $backupPath = Join-Path $backupDir $name
+        Move-Item -Force $dest $backupPath
+        $backedUp.Add($name) | Out-Null
+    }}
+
+    foreach ($name in $stagedNames) {{
+        $dest = Join-Path $installDir $name
+        if (-not (Test-Path $dest)) {{
+            continue
+        }}
+        if ($backedUp.Contains($name)) {{
+            continue
+        }}
+        $backupPath = Join-Path $backupDir $name
+        Move-Item -Force $dest $backupPath
+        $backedUp.Add($name) | Out-Null
+    }}
+
+    foreach ($name in $stagedNames) {{
+        $source = Join-Path $stagingDir $name
+        $dest = Join-Path $installDir $name
+        Move-Item -Force $source $dest
+        $installed.Add($name) | Out-Null
+    }}
+
+    if ($restartAfterUpdate) {{
+        $env:MESH_LLM_SELF_UPDATE_ATTEMPTED = '1'
+        & $exePath @args
+        exit $LASTEXITCODE
+    }}
+    exit 0
+}} catch {{
+    Restore-Backups $backedUp.ToArray() $installed.ToArray()
+    throw
+}} finally {{
+    Remove-Item $workspace -Recurse -Force -ErrorAction SilentlyContinue
+}}
+"#,
+        install_dir = quote(install_dir),
+        workspace = quote(workspace),
+        staging_dir = quote(extracted),
+        backup_dir = quote(backup),
+        exe_path = quote(exe),
+        wait_pid = std::process::id(),
+        restart_after_update = if restart_after_update {
+            "$true"
+        } else {
+            "$false"
+        },
+        managed_json_ps = managed_json_ps,
+        staged_json_ps = staged_json_ps,
+        args_json_ps = args_json_ps
+    ))
+}
+
+#[cfg(not(windows))]
+fn rollback_bundle_replace(
+    install_dir: &Path,
+    backup: &Path,
+    installed: &[String],
+    backed_up: &[String],
+) {
+    for name in installed.iter().rev() {
+        let dest = install_dir.join(name);
+        let _ = std::fs::remove_file(&dest);
+    }
+    for name in backed_up.iter().rev() {
+        let backup_path = backup.join(name);
+        let dest = install_dir.join(name);
+        let _ = std::fs::rename(&backup_path, &dest);
+    }
+}
+
+#[cfg(unix)]
+fn exec_current_binary(exe: &Path) -> Result<()> {
+    let exe_c = CString::new(exe.as_os_str().as_bytes())
+        .context("Executable path contains an unexpected NUL byte")?;
+    let args: Vec<CString> = std::env::args_os()
+        .map(|arg| {
+            CString::new(arg.as_os_str().as_bytes())
+                .context("Argument contains an unexpected NUL byte")
+        })
+        .collect::<Result<_>>()?;
+    let mut argv: Vec<*const libc::c_char> = args.iter().map(|arg| arg.as_ptr()).collect();
+    argv.push(std::ptr::null());
+    let rc = unsafe { libc::execv(exe_c.as_ptr(), argv.as_ptr()) };
+    let errno = std::io::Error::last_os_error();
+    anyhow::ensure!(rc != 0, "execv unexpectedly returned success");
+    Err(errno).context("Failed to restart updated mesh-llm")
+}
+
+#[cfg(not(unix))]
+fn exec_current_binary(_exe: &Path) -> Result<()> {
+    anyhow::bail!("Self-update restart is only supported on Unix")
+}
