+++ crates/mesh-llm-system/src/benchmark/tests.rs
@@ -1,533 +1,3 @@
-use anyhow::{bail, Context, Result};
-use serde::{Deserialize, Serialize};
-use std::path::{Path, PathBuf};
-use std::time::{Duration, Instant};
-
-use crate::hardware::HardwareSurvey;
-
-#[cfg(test)]
-use crate::hardware::GpuFacts;
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct BenchmarkOutput {
-    pub device: String,
-    pub buffer_mb: u32,
-    pub runs: u32,
-    pub p50_gbps: f64,
-    pub p90_gbps: f64,
-    pub compute_tflops_fp32: Option<f64>,
-    pub compute_tflops_fp16: Option<f64>,
-    pub noise_pct: f64,
-    pub runtime_s: f64,
-    pub rated_gbps: Option<f64>,
-    pub rated_estimated: Option<bool>,
-    pub efficiency_pct: Option<f64>,
-    pub bus_width_bits: Option<u32>,
-    pub mem_clock_mhz: Option<u64>,
-    pub gcn_arch: Option<String>,
-    pub hbm: Option<bool>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct GpuBandwidth {
-    pub name: String,
-    pub vram_bytes: u64,
-    pub p50_gbps: f64,
-    pub p90_gbps: f64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub compute_tflops_fp32: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub compute_tflops_fp16: Option<f64>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BenchmarkFingerprint {
-    pub gpus: Vec<GpuBandwidth>, // per-GPU identity + bandwidth, in device order
-    pub is_soc: bool,
-    pub timestamp_secs: u64,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct BenchmarkResult {
-    pub mem_bandwidth_gbps: Vec<f64>,
-    pub compute_tflops_fp32: Option<Vec<f64>>,
-    pub compute_tflops_fp16: Option<Vec<f64>>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct SavedBenchmark {
-    pub path: PathBuf,
-    pub result: BenchmarkResult,
-}
-
-pub const BENCHMARK_TIMEOUT: Duration = Duration::from_secs(25);
-
-/// Normalize `HardwareSurvey.gpu_name` into a per-GPU list of names.
-/// - Splits on ',' and trims whitespace for robustness.
-/// - Expands summarized forms like "8× NVIDIA A100" into 8 identical entries.
-/// - If the expanded list length does not match `gpu_vram.len()` but `gpu_vram` is
-///   non-empty, falls back to assuming all GPUs share the same summarized name and
-///   returns `gpu_vram.len()` copies of it.
-fn per_gpu_names(hw: &HardwareSurvey) -> Vec<String> {
-    let raw = match hw.gpu_name.as_deref() {
-        Some(s) => s.trim(),
-        None => return Vec::new(),
-    };
-
-    if raw.is_empty() {
-        return Vec::new();
-    }
-
-    let mut names: Vec<String> = Vec::new();
-
-    for part in raw.split(',') {
-        let part_trimmed = part.trim();
-        if part_trimmed.is_empty() {
-            continue;
-        }
-
-        // Handle summarized "N× name" form (e.g., "8× NVIDIA A100").
-        if let Some((count_str, name)) = part_trimmed.split_once('×') {
-            if let Ok(count) = count_str.trim().parse::<usize>() {
-                let name_trimmed = name.trim();
-                for _ in 0..count {
-                    names.push(name_trimmed.to_string());
-                }
-                continue;
-            }
-        }
-
-        // Fallback: treat as a single GPU name.
-        names.push(part_trimmed.to_string());
-    }
-
-    if names.len() == hw.gpu_vram.len() || hw.gpu_vram.is_empty() {
-        return names;
-    }
-
-    // As a last resort, assume all GPUs share the same summarized name.
-    let gpu_count = hw.gpu_vram.len();
-    vec![raw.to_string(); gpu_count]
-}
-
-/// Returns true if the current hardware differs from the fingerprint's recorded hardware.
-/// Compares GPU names, VRAM sizes (by index), and the is_soc flag.
-pub fn hardware_changed(fingerprint: &BenchmarkFingerprint, hw: &HardwareSurvey) -> bool {
-    if fingerprint.is_soc != hw.is_soc {
-        return true;
-    }
-
-    let hw_names: Vec<String> = per_gpu_names(hw);
-
-    if fingerprint.gpus.len() != hw_names.len() || fingerprint.gpus.len() != hw.gpu_vram.len() {
-        return true;
-    }
-
-    for (i, cached) in fingerprint.gpus.iter().enumerate() {
-        if cached.name != hw_names[i] || cached.vram_bytes != hw.gpu_vram[i] {
-            return true;
-        }
-    }
-    false
-}
-
-/// Returns the cache-backed benchmark fingerprint path, usually
-/// `~/.cache/mesh-llm/benchmark-fingerprint.json`.
-/// Falls back to `~/.cache` and then the platform temp directory if needed.
-pub fn fingerprint_path() -> PathBuf {
-    dirs::cache_dir()
-        .or_else(|| dirs::home_dir().map(|home| home.join(".cache")))
-        .unwrap_or_else(std::env::temp_dir)
-        .join("mesh-llm")
-        .join("benchmark-fingerprint.json")
-}
-
-fn benchmark_binary_name_for(os: &str, base: &str) -> String {
-    if os == "windows" {
-        format!("{base}.exe")
-    } else {
-        base.to_string()
-    }
-}
-
-fn push_search_dir(dirs: &mut Vec<PathBuf>, dir: PathBuf) {
-    let normalized = dir.canonicalize().unwrap_or(dir);
-    if !dirs.iter().any(|existing| existing == &normalized) {
-        dirs.push(normalized);
-    }
-}
-
-fn benchmark_search_dirs(bin_dir: &Path, exe_dir: Option<&Path>) -> Vec<PathBuf> {
-    let mut dirs = Vec::new();
-    push_search_dir(&mut dirs, bin_dir.to_path_buf());
-
-    if let Some(exe_dir) = exe_dir {
-        push_search_dir(&mut dirs, exe_dir.to_path_buf());
-        push_search_dir(&mut dirs, exe_dir.join("../../target/release"));
-    }
-
-    push_search_dir(&mut dirs, bin_dir.join("../../../target/release"));
-    push_search_dir(&mut dirs, bin_dir.join("../../../../target/release"));
-    dirs
-}
-
-fn detect_benchmark_binary_for_with_exe_dir(
-    os: &str,
-    hw: &HardwareSurvey,
-    bin_dir: &Path,
-    exe_dir: Option<&Path>,
-) -> Option<PathBuf> {
-    if hw.gpu_count == 0 {
-        tracing::debug!("no GPUs detected — skipping benchmark");
-        return None;
-    }
-
-    let gpu_upper = hw.gpu_name.as_deref().unwrap_or("").to_uppercase();
-
-    let candidate_name = if os == "macos" && hw.is_soc {
-        benchmark_binary_name_for(os, "membench-fingerprint")
-    } else if os == "linux" || os == "windows" {
-        if gpu_upper.contains("NVIDIA") {
-            benchmark_binary_name_for(os, "membench-fingerprint-cuda")
-        } else if gpu_upper.contains("AMD") || gpu_upper.contains("RADEON") {
-            benchmark_binary_name_for(os, "membench-fingerprint-hip")
-        } else if gpu_upper.contains("INTEL") || gpu_upper.contains("ARC") {
-            tracing::info!("Intel Arc benchmark is unvalidated — results may be inaccurate");
-            benchmark_binary_name_for(os, "membench-fingerprint-intel")
-        } else if os == "linux" && hw.is_soc {
-            tracing::warn!("Jetson benchmark is unvalidated for ARM CUDA — attempting");
-            benchmark_binary_name_for(os, "membench-fingerprint-cuda")
-        } else {
-            tracing::warn!(
-                "could not identify benchmark binary for this GPU platform: {:?}",
-                hw.gpu_name
-            );
-            return None;
-        }
-    } else {
-        tracing::warn!(
-            "could not identify benchmark binary for this GPU platform: {:?}",
-            hw.gpu_name
-        );
-        return None;
-    };
-
-    for search_dir in benchmark_search_dirs(bin_dir, exe_dir) {
-        let candidate = search_dir.join(&candidate_name);
-        if candidate.exists() {
-            return Some(candidate);
-        }
-    }
-
-    tracing::warn!(
-        "{candidate_name} not found in benchmark search dirs: {:?}",
-        benchmark_search_dirs(bin_dir, exe_dir)
-    );
-    None
-}
-
-/// Load a `BenchmarkFingerprint` from disk.  Returns `None` on any error.
-pub fn load_fingerprint(path: &Path) -> Option<BenchmarkFingerprint> {
-    let content = std::fs::read_to_string(path).ok()?;
-    serde_json::from_str(&content).ok()
-}
-
-/// Atomically write a `BenchmarkFingerprint` to disk.
-/// Uses a `.json.tmp` staging file + rename for crash safety.
-/// Logs a warning on failure — never panics.
-pub fn save_fingerprint(path: &Path, fp: &BenchmarkFingerprint) {
-    if let Err(err) = try_save_fingerprint(path, fp) {
-        tracing::warn!("benchmark: failed to persist fingerprint: {err}");
-    }
-}
-
-pub fn try_save_fingerprint(path: &Path, fp: &BenchmarkFingerprint) -> Result<()> {
-    let tmp = path.with_extension("json.tmp");
-
-    std::fs::create_dir_all(path.parent().unwrap_or_else(|| Path::new(".")))
-        .with_context(|| format!("failed to create cache dir for {}", path.display()))?;
-
-    let json =
-        serde_json::to_string_pretty(fp).context("failed to serialize benchmark fingerprint")?;
-
-    std::fs::write(&tmp, &json)
-        .with_context(|| format!("failed to write temporary fingerprint {}", tmp.display()))?;
-
-    // On Windows, `rename` fails if the destination already exists.
-    // Remove the destination first there; on Unix the rename stays atomic.
-    #[cfg(windows)]
-    if path.exists() {
-        std::fs::remove_file(path)
-            .with_context(|| format!("failed to remove existing fingerprint {}", path.display()))?;
-    }
-
-    if let Err(e) = std::fs::rename(&tmp, path) {
-        let _ = std::fs::remove_file(&tmp);
-        return Err(e).with_context(|| {
-            format!(
-                "failed to rename fingerprint into place at {}",
-                path.display()
-            )
-        });
-    }
-
-    Ok(())
-}
-
-/// Determine which benchmark binary to use for the current hardware platform.
-///
-/// Returns `None` (soft failure) if:
-/// - No GPUs are present
-/// - The binary is not found on disk
-/// - The platform/GPU combination is unrecognised
-///
-/// Never panics or hard-fails with `ensure!`.
-pub fn detect_benchmark_binary(hw: &HardwareSurvey, bin_dir: &Path) -> Option<PathBuf> {
-    let exe_dir = std::env::current_exe()
-        .ok()
-        .and_then(|exe_path| exe_path.parent().map(Path::to_path_buf));
-    detect_benchmark_binary_for_with_exe_dir(std::env::consts::OS, hw, bin_dir, exe_dir.as_deref())
-}
-
-/// Parse raw stdout bytes from a benchmark run into a vec of per-device outputs.
-///
-/// Expects a JSON array of [`BenchmarkOutput`].  Returns `None` on any parse
-/// failure or if the device list is empty.
-pub fn parse_benchmark_output(stdout: &[u8]) -> Option<Vec<BenchmarkOutput>> {
-    match serde_json::from_slice::<Vec<BenchmarkOutput>>(stdout) {
-        Ok(outputs) if !outputs.is_empty() => Some(outputs),
-        Ok(_) => {
-            tracing::debug!("benchmark returned empty device list");
-            None
-        }
-        Err(err) => {
-            if let Ok(val) = serde_json::from_slice::<serde_json::Value>(stdout) {
-                if let Some(msg) = val.get("error").and_then(|v| v.as_str()) {
-                    tracing::warn!("benchmark reported error: {msg}");
-                    return None;
-                }
-            }
-            tracing::warn!("failed to parse benchmark output: {err}");
-            None
-        }
-    }
-}
-
-/// Run the benchmark binary synchronously and return per-device outputs.
-///
-/// Spawns the binary as a subprocess and polls for completion up to `timeout`.
-/// If the process exceeds the timeout, it is killed to avoid zombie processes.
-///
-/// Designed to be called inside `tokio::task::spawn_blocking` — never `async`.
-pub fn run_benchmark(binary: &Path, timeout: Duration) -> Option<Vec<BenchmarkOutput>> {
-    use std::io::Read;
-
-    let mut child = match std::process::Command::new(binary)
-        .arg("--json")
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()
-    {
-        Ok(c) => c,
-        Err(e) => {
-            tracing::error!("failed to spawn {binary:?}: {e}");
-            return None;
-        }
-    };
-
-    let deadline = Instant::now() + timeout;
-    let status = loop {
-        match child.try_wait() {
-            Ok(Some(status)) => break status,
-            Ok(None) => {
-                if Instant::now() >= deadline {
-                    tracing::warn!("benchmark timed out after {timeout:?}, killing subprocess");
-                    let _ = child.kill();
-                    let _ = child.wait();
-                    return None;
-                }
-                std::thread::sleep(Duration::from_millis(100));
-            }
-            Err(e) => {
-                tracing::error!("error waiting for benchmark: {e}");
-                let _ = child.kill();
-                let _ = child.wait();
-                return None;
-            }
-        }
-    };
-
-    if !status.success() {
-        tracing::warn!("benchmark exited with {:?}", status);
-        return None;
-    }
-
-    let mut stdout_bytes = Vec::new();
-    if let Some(mut pipe) = child.stdout.take() {
-        let _ = pipe.read_to_end(&mut stdout_bytes);
-    }
-    parse_benchmark_output(&stdout_bytes)
-}
-
-/// Load a cached fingerprint if hardware is unchanged, otherwise run the
-/// benchmark binary and persist the result.
-///
-/// Not `async` — intended for use inside `tokio::task::spawn_blocking`.
-pub fn run_or_load(
-    hw: &HardwareSurvey,
-    bin_dir: &Path,
-    timeout: Duration,
-) -> Option<BenchmarkResult> {
-    let path = fingerprint_path();
-
-    // Cache-hit path
-    if let Some(ref cached) = load_fingerprint(&path) {
-        if !hardware_changed(cached, hw) {
-            let mem_bandwidth: Vec<f64> = cached.gpus.iter().map(|g| g.p90_gbps).collect();
-            let compute_tflops_fp32 = cached
-                .gpus
-                .iter()
-                .map(|g| g.compute_tflops_fp32)
-                .collect::<Option<Vec<f64>>>();
-            let compute_tflops_fp16 = cached
-                .gpus
-                .iter()
-                .map(|g| g.compute_tflops_fp16)
-                .collect::<Option<Vec<f64>>>();
-            let result = BenchmarkResult {
-                mem_bandwidth_gbps: mem_bandwidth,
-                compute_tflops_fp32,
-                compute_tflops_fp16,
-            };
-            tracing::info!(
-                "Using cached bandwidth fingerprint: {} GPUs",
-                result.mem_bandwidth_gbps.len()
-            );
-            return Some(result);
-        }
-    }
-
-    tracing::info!("Hardware changed or no cache — running memory bandwidth benchmark");
-
-    let binary = detect_benchmark_binary(hw, bin_dir)?;
-    let outputs = run_benchmark(&binary, timeout)?;
-
-    let (gpus, result) = build_benchmark_result(hw, &outputs);
-
-    let fingerprint = BenchmarkFingerprint {
-        gpus,
-        is_soc: hw.is_soc,
-        timestamp_secs: std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs(),
-    };
-
-    save_fingerprint(&path, &fingerprint);
-    Some(result)
-}
-
-pub fn run_and_save(
-    hw: &HardwareSurvey,
-    bin_dir: &Path,
-    timeout: Duration,
-) -> Result<SavedBenchmark> {
-    run_and_save_to_path(hw, bin_dir, timeout, &fingerprint_path())
-}
-
-fn run_and_save_to_path(
-    hw: &HardwareSurvey,
-    bin_dir: &Path,
-    timeout: Duration,
-    path: &Path,
-) -> Result<SavedBenchmark> {
-    if hw.gpu_count == 0 {
-        bail!("no GPUs detected on this node");
-    }
-
-    let binary = detect_benchmark_binary(hw, bin_dir).with_context(|| {
-        format!(
-            "no supported benchmark binary found for detected GPU platform {:?}",
-            hw.gpu_name
-        )
-    })?;
-
-    let outputs = run_benchmark(&binary, timeout)
-        .with_context(|| format!("benchmark run failed for {}", binary.display()))?;
-
-    let result = save_result_from_outputs(path, hw, &outputs)?;
-    Ok(SavedBenchmark {
-        path: path.to_path_buf(),
-        result,
-    })
-}
-
-fn save_result_from_outputs(
-    path: &Path,
-    hw: &HardwareSurvey,
-    outputs: &[BenchmarkOutput],
-) -> Result<BenchmarkResult> {
-    let (gpus, result) = build_benchmark_result(hw, outputs);
-
-    let fingerprint = BenchmarkFingerprint {
-        gpus,
-        is_soc: hw.is_soc,
-        timestamp_secs: std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs(),
-    };
-
-    try_save_fingerprint(path, &fingerprint)?;
-    Ok(result)
-}
-
-fn build_benchmark_result(
-    hw: &HardwareSurvey,
-    outputs: &[BenchmarkOutput],
-) -> (Vec<GpuBandwidth>, BenchmarkResult) {
-    let hw_names = per_gpu_names(hw);
-
-    let count = outputs
-        .len()
-        .min(hw.gpu_vram.len())
-        .min(if hw_names.is_empty() {
-            usize::MAX
-        } else {
-            hw_names.len()
-        });
-
-    let gpus: Vec<GpuBandwidth> = (0..count)
-        .map(|i| GpuBandwidth {
-            name: hw_names.get(i).cloned().unwrap_or_default(),
-            vram_bytes: hw.gpu_vram.get(i).copied().unwrap_or(0),
-            p50_gbps: outputs[i].p50_gbps,
-            p90_gbps: outputs[i].p90_gbps,
-            compute_tflops_fp32: outputs[i].compute_tflops_fp32,
-            compute_tflops_fp16: outputs[i].compute_tflops_fp16,
-        })
-        .collect();
-
-    let mem_bandwidth_gbps = gpus.iter().map(|g| g.p90_gbps).collect();
-    let compute_tflops_fp32 = gpus
-        .iter()
-        .map(|g| g.compute_tflops_fp32)
-        .collect::<Option<Vec<f64>>>();
-    let compute_tflops_fp16 = gpus
-        .iter()
-        .map(|g| g.compute_tflops_fp16)
-        .collect::<Option<Vec<f64>>>();
-
-    (
-        gpus,
-        BenchmarkResult {
-            mem_bandwidth_gbps,
-            compute_tflops_fp32,
-            compute_tflops_fp16,
-        },
-    )
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
